Refactor integer and string functions
